add tests for effect.console and effect.files
